Significantly change User model.
Since we're launching Oxford only, this commit adds a `college' field.
It also adds `first_name' and `last_name' fields, which makes editing
more personal.

This also adds in auto-stripping atributes, via a gem. The `email',
`first_name', and `last_name' attributes are all stripped of whitespace.

This also adds a capitalization module, which capitalizes first and last
names. This is a seperate module written by myself. Since this is in
`lib', the whole folder has been added to the autoload_path.

These new fields are now required in the user model. Thus, the
registraion edit and new pages have also changed. Along with them,
autofilling of the `first_name', `last_name', and `college' fields is
done via e-mail.