Remove Oxford email requirement.
Previously, Oxford emails were required, both in the model, and on the
front-end sign-up page [1]. This relaxes that requirement, for demos.

This also removes messages on the sign-in and edit-account pages that
states an Oxford email is required.

1. Thhey were never required on the front-end edit-account page, since
admins 1) use that page, and 2) aren't required to have `.ox.ac.uk'
emails.